backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.0
pydantic==1.10.7
python-magic==0.4.27
httpx==0.27.0
firebase-admin==6.5.0